split functions into lib
